requirements.txt
# Core dependencies
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5

# Vector database
chromadb>=0.4.0

# Text processing
langchain-text-splitters>=0.0.1

# OpenAI
openai>=1.0.0

# HuggingFace (for open-source embedding models)
langchain-huggingface>=0.0.1
sentence-transformers>=2.2.0
torch>=2.0.0

# Web UI
streamlit>=1.30.0

# Evaluation & analysis
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0